requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
gunicorn>=21.2.0
uvloop>=0.19.0
httptools>=0.6.1

# Database
sqlalchemy>=2.0.23
psycopg[binary]>=3.1.13
alembic>=1.12.1

# Data Validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Authentication & Security
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Image Processing
Pillow>=10.1.0

# Email
aiosmtplib>=3.0.1

# Rate Limiting
slowapi>=0.1.9

# Redis
redis>=5.0.1
hiredis>=2.2.3

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
httpx>=0.25.2

# Logging
python-json-logger>=2.0.7
